🎨 Move theme middleware to theme module
